fix(nestjs-signed-url): align interface ownership
