+++ packages/nestjs-signed-url/src/gcs/index.ts
@@ -1,3 +1,4 @@
 export * from './constants.js'
 export * from './gateway.js'
+export type * from './interfaces.js'
 export * from './signer.js'
