+++ packages/nestjs-signed-url/src/interfaces.ts
