+++ packages/nestjs-signed-url/src/tests/gcs.module.fixture.ts
@@ -1,23 +1,24 @@
 import type { Storage }       from '@google-cloud/storage'
 import type { DynamicModule } from '@nestjs/common'
-
-export const TESTING_GCS_STORAGE_FACTORY = Symbol('testing-gcs-storage-factory')
-
-export type TestingGcsStorageFactory = {
-  create: () => Storage
-}
+import type { TestingGcsStorageFactory } from './gcs.module.interfaces.js'
 
 class TestingGcsClientModule {}
 
-export const createTestingGcsClientModule = (storage: Storage): DynamicModule => ({
-  module: TestingGcsClientModule,
-  providers: [
-    {
-      provide: TESTING_GCS_STORAGE_FACTORY,
-      useValue: {
-        create: (): Storage => storage,
+export const TESTING_GCS_STORAGE_FACTORY = Symbol('testing-gcs-storage-factory')
+
+export const createTestingGcsClientModule = (storage: Storage): DynamicModule => {
+  const storageFactory: TestingGcsStorageFactory = {
+    create: (): Storage => storage,
+  }
+
+  return {
+    module: TestingGcsClientModule,
+    providers: [
+      {
+        provide: TESTING_GCS_STORAGE_FACTORY,
+        useValue: storageFactory,
       },
-    },
-  ],
-  exports: [TESTING_GCS_STORAGE_FACTORY],
-})
+    ],
+    exports: [TESTING_GCS_STORAGE_FACTORY],
+  }
+}
