+++ packages/nestjs-signed-url/src/tests/gcs.module.interfaces.ts
@@ -1,0 +1,5 @@
+import type { Storage } from '@google-cloud/storage'
+
+export interface TestingGcsStorageFactory {
+  create: () => Storage
+}
